[예약] 예약현황 스타일 수정, rt 오류 수정

+++ src/api/userApi.jsx
@@ -66,6 +66,7 @@ const fetchUserStateQuery = async (param, search) => {
 
 export const logoutUser = async (id, rt) => {
 	try {
+		if (id === 0 || rt === null) return;
 		const response = await fetch(`${process.env.REACT_APP_API_URL}/auth/logout?userId=${id}`, {
 			method: "DELETE",
 			headers: {
